Move Python comms code to new package
